requirements.txt
pyyaml>=3.12
tornado>=4.4.2
requests>=1.3.1